V1 estable funcionando con los reportes y todo en https correjido todo el 30 de octubre 2016 a las 12medio dia colombia

+++ app/Http/Controllers/configuracion/administradorController.php
@@ -92,8 +92,17 @@ class administradorController extends Controller
 		}
 		else
 		{
-			$datos->delete();
+			$validar=administrador::where('tipo','=','superAdmin')->where('idadministrador','=',Auth::user()->id)->count();
+			if($validar>0)
+			{
+				$datos->delete();
 			$this->show();
+			}
+			else
+			{
+				echo "2";
+			}
+			
 		}
 	}
 }
